+++ libs/cardano-ledger-core/src/Cardano/Ledger/TxIn.hs
@@ -23,15 +23,20 @@ module Cardano.Ledger.TxIn (
 )
 where
 
+import Cardano.Crypto.Hash.Class (hashToTextAsHex)
 import Cardano.HeapWords (HeapWords (..))
 import qualified Cardano.HeapWords as HW
 import Cardano.Ledger.BaseTypes (TxIx (..), mkTxIxPartial)
 import Cardano.Ledger.Binary (DecCBOR (decCBOR), EncCBOR (..), decodeRecordNamed, encodeListLen)
-import qualified Cardano.Ledger.Crypto as CC
+import Cardano.Ledger.Crypto
 import Cardano.Ledger.Hashes (EraIndependentTxBody)
-import Cardano.Ledger.SafeHash (SafeHash)
+import Cardano.Ledger.SafeHash (SafeHash, extractHash)
 import Cardano.Ledger.TreeDiff (ToExpr)
 import Control.DeepSeq (NFData)
+import Data.Aeson (FromJSON, ToJSON (..))
+import Data.Aeson.Types (ToJSONKey (..), toJSONKeyText)
+import Data.Text (Text)
+import qualified Data.Text as Text
 import GHC.Generics (Generic)
 import GHC.Stack (HasCallStack)
 import NoThunks.Class (NoThunks (..))
@@ -49,19 +54,32 @@ import NoThunks.Class (NoThunks (..))
 -- | A unique ID of a transaction, which is computable from the transaction.
 newtype TxId c = TxId {_unTxId :: SafeHash c EraIndependentTxBody}
   deriving (Show, Eq, Ord, Generic)
-  deriving newtype (NoThunks)
+  deriving newtype (NoThunks, ToJSON, FromJSON)
 
-deriving newtype instance CC.Crypto c => HeapWords (TxId c)
+deriving newtype instance Crypto c => HeapWords (TxId c)
 
-deriving newtype instance CC.Crypto c => EncCBOR (TxId c)
+deriving newtype instance Crypto c => EncCBOR (TxId c)
 
-deriving newtype instance CC.Crypto c => DecCBOR (TxId c)
+deriving newtype instance Crypto c => DecCBOR (TxId c)
 
-deriving newtype instance CC.Crypto c => NFData (TxId c)
+deriving newtype instance Crypto c => NFData (TxId c)
 
-instance CC.Crypto c => HeapWords (TxIn c) where
+instance Crypto c => HeapWords (TxIn c) where
   heapWords (TxIn txId _) =
     2 + HW.heapWords txId + 1 {- txIx -}
+
+instance Crypto c => ToJSON (TxIn c) where
+  toJSON = toJSON . txInToText
+  toEncoding = toEncoding . txInToText
+
+instance Crypto c => ToJSONKey (TxIn c) where
+  toJSONKey = toJSONKeyText txInToText
+
+txInToText :: TxIn c -> Text
+txInToText (TxIn (TxId txidHash) ix) =
+  hashToTextAsHex (extractHash txidHash)
+    <> Text.pack "#"
+    <> Text.pack (show ix)
 
 -- | The input of a UTxO.
 data TxIn c = TxIn !(TxId c) {-# UNPACK #-} !TxIx
@@ -78,17 +96,17 @@ deriving instance Ord (TxIn c)
 
 deriving instance Show (TxIn c)
 
-deriving instance CC.Crypto c => NFData (TxIn c)
+deriving instance Crypto c => NFData (TxIn c)
 
 instance NoThunks (TxIn c)
 
-instance CC.Crypto c => EncCBOR (TxIn c) where
+instance Crypto c => EncCBOR (TxIn c) where
   encCBOR (TxIn txId index) =
     encodeListLen 2
       <> encCBOR txId
       <> encCBOR index
 
-instance CC.Crypto c => DecCBOR (TxIn c) where
+instance Crypto c => DecCBOR (TxIn c) where
   decCBOR =
     decodeRecordNamed
       "TxIn"