+++ eras/allegra/impl/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.0.0.1
 
+### `testlib`
+
 * Consolidate all `Arbitrary` instances from the test package to under a new `testlib`. #3285
 
 ## 1.0.0.0