+++ eras/alonzo/impl/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Version history for `cardano-ledger-alonzo`
 
-## 1.0.0.1
+## 1.1.0.0
+
+* Add `ToJSON` instance for `AlonzoTxOut`, `AlonzoScript` and `Datum`
+* Add `ToJSON` instance for `AlonzoPParams StrictMaybe`
+
+###`testlib`
 
 * Consolidate all `Arbitrary` instances from the test package to under a new `testlib`. #3285