+++ eras/babbage/impl/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Version history for `cardano-ledger-babbage`
 
-## 1.0.0.1
+## 1.1.0.0
+
+* Add `ToJSON` instance for `BabbageTxOut`.
+* Add `ToJSON` instance for `BabbagePParams Identity` and `BabbagePParams StrictMaybe`
+
+###`testlib`
 
 * Consolidate all `Arbitrary` instances from the test package to under a new `testlib`. #3285