+++ eras/mary/impl/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Version history for `cardano-ledger-mary`
 
-## 1.0.0.1
+## 1.1.0.0
+
+* Addition of `ToJSON` instances for `AssetName`, `PolicyID`, `MultiAsset` and `MaryValue`.
+
+### `testlib`
 
 * Consolidate all `Arbitrary` instances from the test package to under a new `testlib`. #3285