+++ libs/cardano-ledger-binary/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Version history for `cardano-ledger-binary`
 
+## 1.0.1.0
+
+* Add `ToJSON`/`FromJSON` instances for `Version`
+* Add `decodeFullFromHexText` and `serializeAsHexText`
+
+### `testlib`
+
+* Add `Arbitrary` instance for `Term`
+
 ## 1.0.0.0
 
 * First version. Released on an unsuspecting world.