+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Version history for `cardano-ledger-core`
 
+## 1.1.0.0
+
+* Add `ToJSON (PParamsHKD f era)` superclass constraints for `EraPParams`.
+* Add `ToJSON (TxOut era)` superclass constraints for `EraTxOut`.
+* Add superclass constraints for `Val`:
+  * `NoThunks`, `EncCBOR`, `DecCBOR`, `ToJSON`, `NFData`, `Show`
+  * `EncCBOR (CompactForm t)`, `DecCBOR (CompactForm t)`
+* Add `ToJSON`/`FromJSON` instances for `CompactForm Coin`, `SafeHash` and `TxId`
+* Add `ToJSON` instances for:
+  * `Ptr`, `CertIx`, `TxIx`
+  * `Trip` and `UMap`
+  * `DeltaCoin` and `CompactForm DeltaCoin`
+  * `InstantaneousRewards`, `FutureGenDeleg`, `PState`, `DState` and `DPState`.
+  * `UTxO` and `TxIn`
+  * `Stake`, `SnapShot`, `SnapShots`, `PoolDistr` and `IndividualPoolStake`
+  * `Reward` and `RewardType`
+
 ## 1.0.0.0
 
 * First properly versioned release.